ahss_backend/ahss_backend/requirements.txt
Django==2.2.20
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
mccabe==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
